# Make quick-sqlite 2x quicker?