--- a/cpp/sqliteBridge.h
+++ b/cpp/sqliteBridge.h
@@ -25,6 +25,8 @@ SQLiteOPResult sqliteDetachDb(string const mainDBName, string const alias);
 
 SQLiteOPResult sqliteExecute(string const dbName, string const &query, vector<QuickValue> *values, vector<map<string, QuickValue>> *result, vector<QuickColumnMetadata> *metadata);
 
+SQLiteOPResult sqliteExecute2(jsi::Runtime &rt, string const dbName, string const &query, vector<QuickValue> *values, bool const returnArrays, vector<jsi::Object> *result, vector<QuickColumnMetadata> *metadata);
+
 SequelLiteralUpdateResult sqliteExecuteLiteral(string const dbName, string const &query);
 
 void sqliteCloseAll();