@ospfranco 2023-10-31
The wrapper QuickValue is just a convenient way to encapsulate any type of values and return them to the calling JSI function to be later returned to the JS context. Memory allocation probably takes some time, so you are right, getting rid should decrease memory consumption.