@mjmasn 2023-10-31
Thanks @ospfranco, appreciate the quick reply! I figured that'd be the reasoning but that's positive in a way, I wanted to make sure there wasn't a hard technical reason why it would be necessary.

I guess we'll patch this on our end for now as we do need to prioritise perf, but I still think there's value in exploring performance improvements more generally.

Will leave this open for comments anyway