@ospfranco 2023-11-03
I just took another look into your code and I realized you only modified the sync operations. You might be lacking some context here. 

Indeed it is no problem to create more JSI objects on sync operations and you will gain performance because you are halting the JavaScript context. There you can pass and create jsi::XXX values as the query executes because no other JS is running.

The `QuickValue` abstraction was there to allow the package to run async queries in a different thread, and then return the results to the calling function purely in C++ scalars, which then needs to [schedule blocking time in the JS context again](https://github.com/margelo/react-native-quick-sqlite/blob/main/cpp/bindings.cpp#L237) in order to create JS values again. If you don't do this, then you will crash JS because it might be busy in the middle of another operation. This is just a limitation of interacting between native code and JS. Although Async operations might be a little slower, they do not block your React Native app from running and should be the go-to way for such large operations. At the end of the day, one could have both abstractions, but I developed the package in my free time and it is too much of a time sink to maintain two different implementations.

You also posted some memory crashes, which will depend on the device. You tested on some somewhat older devices, but there are a lot of smaller Android devices out there with very little memory and such large datasets will almost always crash on those. Even though I tried to make the fastest possible package, I still do not encourage running such large data sets on mobile devices.

In any case, thank you for posting some metrics! I also took a look at the codebase again after many months and really enjoyed spelunking once again :)