@vhakulinen 2023-11-03
This might be a bit off topic, but inspired by not too dissimilar reasons I started to work on new sqlite bindings for JSI with a different API approach. The code can be found here: https://github.com/vhakulinen/sqlite-jsi. I would like to crunch it out to a complete package for react native, since but I don't have the time for it, perhaps the code can be a source of inspiration others.

One thing you could try with the values (i.e. `QuickValue`) is to turn them into hosted objects. This way you might be able to avoid copying. I.e. instead of
* sqlite **_copy_** QuickValue **_copy_** runtime

you would have

* sqlite **_copy_** QuickValue (hosted object) **_on demand_** runtime (perhaps without copying through `ArrayBuffer`).

I just realized this and haven't tried it in `sqlite-jsi`.

Another thing I suspect that might be worth of a closer look is the connection initialization. Currently `react-native-quick-sqlite` [forces](https://github.com/margelo/react-native-quick-sqlite/blob/be9235eef7d892ed46177f4d4031cc1a9af723ad/cpp/sqliteBridge.cpp#L87) serialized threading mode. Reworking the code to work through [other modes](https://www.sqlite.org/threadsafe.html) could improve performance on some multithreaded workloads. With `sqlite-jsi`, I limited background threads to one per connection (and put a mutex in front of the connection) and leave it up to the API consumer to open more connections (i.e. connection pool) for parallel work. Having per connection blocking transaction in this scenario is a bit challenging, but doable.

And a third thing, preserve and reuse the prepared statements. Currently `react-native-quick-sqlite` doesn't allow this and throws the statements away after each API call.


Lastly, thank you @ospfranco for creating this library. I has helped us immensely.