@vhakulinen 2023-11-04
Once you have thousands of items loaded from the database and displayed in, say, scrolled view the cost of copying _everything_ upfront starts to matter.

And the threading issue is likely that the code is currently sharing the same connection object across threads which is a big no-no without mutexes (which the serialized mode adds).

@mjmasn one additional thing for the string performance: javascript handles strings in UTF-16, while JSI (from what I've seen) and sqlite uses UTF-8. Converting UTF-8 <-> UTF-16 has some cost. If you're loading a ton of text from sqlite that you're not displaying immediately to the user (or use otherwise), you could try to query the texts as BLOBs, load them to a array buffer and then, once required, turn the array buffer into a string.