@ospfranco 2023-11-04
> Once you have thousands of items loaded from the database and displayed in, say, scrolled view the cost of copying _everything_ upfront starts to matter.

Yeah, you are right. I cannot stop people from squeezing every drop of performance out of the devices.

> And the threading issue is likely that the code is currently sharing the same connection object across threads which is a big no-no without mutexes (which the serialized mode adds).

Yeah, I think that was the problem. I think at some point I tried multiple connections but the code gets messy fast.

> @mjmasn one additional thing for the string performance: javascript handles strings in UTF-16, while JSI (from what I've seen) and sqlite uses UTF-8. Converting UTF-8 <-> UTF-16 has some cost. If you're loading a ton of text from sqlite that you're not displaying immediately to the user (or use otherwise), you could try to query the texts as BLOBs, load them to a array buffer and then, once required, turn the array buffer into a string.

JavaScript is retarded. Indeed internally every string is UTF-16 and there are a lot of conversions between UTF-8 and UTF-16 even when it interacts within itself.

Actually, you guys inspired me to try to hack at this again and see what can be done to really squeeze the last bit of performance. Instead of using the QuickValue struct one can use std::any, which should use a lot less memory allocation. I also implemented a new version of a SmartHostObject (over a simpler previous implementation, the new version that uses an unordered_map internally which means faster access to fields) and this is what I got.

I created a sample database with 300k records. A mixture of strings, ints, and floats. Then queried all of them at once.

Using current implementation:

<img src="https://github.com/margelo/react-native-quick-sqlite/assets/1634213/3d9ecfd2-c8b3-46f9-894d-2e9209a294ec" width="200" />

New implementation:

<img src="https://github.com/margelo/react-native-quick-sqlite/assets/1634213/2a014739-477e-4c76-a701-556ef07a225b" width="200" />