@ospfranco 2023-11-06
Created a new package. Mostly API compatible, you will need to relocate your DB or specify a path. But it is quite faster when loading large sets of data.

https://github.com/OP-Engineering/op-sqlite