@mjmasn 2023-11-07
Just catching up on this thread, could not have hoped for a better set of replies ❤️ 

In my initial tests I'm seeing a 3x perf boost on SELECTs using your new package @ospfranco, really appreciate you taking the time to look into this.