@ospfranco 2023-11-07
cheers! access are a bit slower, so if you iterate over all of the data it will probably be slower, but I guess it is better when used in flatlists and what not.