@ospfranco 2023-11-07
I just published a new version. I've now gotten Android to under a second and iOS to ~500ms to load the 300k records. @mjmasn give it a try.